graph/encoding/dot: reorganise files
